requirements.txt
discord.py
asyncio
matplotlib
requests
colorthief
aiohttp
asyncpg
uvloop
aiofiles
pillow
jishaku
wikipedia-api
pytz
typing_extensions
dateparser
pygit2
psutil
python-dotenv
lru-dict
git+https://github.com/blanketsucks/aiospotify